+++ lnbits/core/services/assets.py
@@ -1,5 +1,6 @@
 import base64
 import io
+from urllib.parse import quote
 from uuid import uuid4
 
 import filetype
@@ -23,6 +24,25 @@ IMAGE_MIME_TYPE_ALIASES = {
 PIL_IMAGE_FORMAT_MIME_TYPES = {
     "JPEG": "image/jpeg",
     "PNG": "image/png",
+}
+INLINE_ASSET_MIME_TYPES = {
+    "image/heic",
+    "image/heics",
+    "image/heif",
+    "image/jpeg",
+    "image/png",
+}
+ASSET_SECURITY_HEADERS = {
+    "X-Content-Type-Options": "nosniff",
+    "Content-Security-Policy": (
+        "sandbox; default-src 'none'; script-src 'none'; "
+        "object-src 'none'; base-uri 'none'"
+    ),
+}
+THUMBNAIL_FORMAT_MIME_TYPES = {
+    "jpg": "image/jpeg",
+    "jpeg": "image/jpeg",
+    "png": "image/png",
 }
 
 
@@ -85,6 +105,23 @@ async def create_user_asset(user_id: str, file: UploadFile, is_public: bool) -> 
 def normalize_asset_mime_type(content_type: str) -> str:
     content_type = content_type.split(";", 1)[0].strip().lower()
     return IMAGE_MIME_TYPE_ALIASES.get(content_type, content_type)
+
+
+def normalize_media_type(media_type: str) -> str:
+    return media_type.split(";", 1)[0].strip().lower() or "application/octet-stream"
+
+
+def thumbnail_media_type() -> str:
+    thumbnail_format = (settings.lnbits_asset_thumbnail_format or "png").strip().lower()
+    return THUMBNAIL_FORMAT_MIME_TYPES.get(thumbnail_format, "application/octet-stream")
+
+
+def content_disposition(disposition: str, filename: str) -> str:
+    safe_filename = filename or "unnamed"
+    quoted_filename = quote(safe_filename, safe="")
+    if quoted_filename == safe_filename:
+        return f'{disposition}; filename="{safe_filename}"'
+    return f"{disposition}; filename*=utf-8''{quoted_filename}"
 
 
 def allowed_asset_mime_types() -> set[str]:
